Serialization fix & Execution manager basic structure
Fixed include of serialization (instead of header-only, created cpp file for all definitions).

Also, created class for Execution manager and added required methods.
