+++ applications/ExecutionManager/src/execution_manager.cpp
@@ -1,0 +1,20 @@
+#include "execution_manager.hpp"
+
+namespace ExecutionManager {
+
+int ExecutionManager::loadFileNamesInDir(string path, vector<string> &fileNames)
+{
+    return 0;
+}
+
+vector<ApplicationManifest> ExecutionManager::processManifests()
+{
+    return {};
+}
+
+applicationId ExecutionManager::startApplication(ApplicationManifest &manifest)
+{
+    return "";
+}
+
+} // namespace ExecutionManager
