+++ applications/ExecutionManager/src/manifests.cpp
@@ -1,5 +1,4 @@
-#ifndef JSON_CONVERTERS_HPP
-#define JSON_CONVERTERS_HPP
+#include "manifests.hpp"
 
 #include <json.hpp>
 #include <application_state_client.h>
@@ -10,7 +9,92 @@ using nlohmann::json;
 namespace ExecutionManager
 {
 
-// MachineStates serialization & deserialization
+void MachineManifest::init()
+{
+  loadHwConf();
+  loadNetworkConf();
+  states = {MachineStates::kRunning};
+  adaptiveModules = {};
+}
+
+void MachineManifest::loadNetworkConf()
+{
+  ifaddrs *ifa;
+  getifaddrs(&ifa);
+
+  for (ifaddrs *it = ifa; it != nullptr; it = it->ifa_next)
+  {
+    InterfaceConf interfaceConf;
+    interfaceConf.ifa_name = string{it->ifa_name};
+    auto family = it->ifa_addr->sa_family;
+
+    interfaceConf.family =
+        (family == AF_INET)
+            ? "AF_INET"
+            : (family == AF_INET6)
+                  ? "AF_INET6"
+                  : (family == AF_PACKET) ? "AF_PACKET" : "Unknown";
+
+    if (family == AF_INET || family == AF_INET6)
+    {
+      int err = getnameinfo(it->ifa_addr,
+                            (family == AF_INET) ? sizeof(struct sockaddr_in)
+                                                : sizeof(struct sockaddr_in6),
+                            interfaceConf.host, NI_MAXHOST, nullptr, 0,
+                            NI_NUMERICHOST);
+      if (err != 0)
+      {
+        std::cerr << "getnameinfo() failed " << gai_strerror(err) << std::endl;
+        continue;
+      }
+    }
+
+    network.push_back(interfaceConf);
+  }
+
+  freeifaddrs(ifa);
+}
+
+void MachineManifest::loadHwConf()
+{
+  // load ram available
+  struct sysinfo info;
+  if (sysinfo(&info))
+  {
+    std::cout << "can't load ram info\n";
+    return;
+  }
+  hwConf.ram = info.freeram;
+
+  // load available cpu power
+  size_t prevIdleTime = 0;
+  size_t prevTotalTime = 0;
+
+#if UNIT_TEST
+  ifstream procStat("../test/cpu-mockfile");
+#else
+  ifstream procStat("/proc/stat");
+#endif
+  procStat.ignore(5, ' ');  // skip cpu prefix
+  vector<size_t> times;
+  for (size_t time; procStat >> time; times.push_back(time))
+    ;
+
+  if (times.size() < 4)
+  {
+    std::cout << "Error obtaining cpu data\n";
+  }
+
+  size_t idleTime = times[3];
+  size_t totalTime = std::accumulate(times.begin(), times.end(), 0);
+
+  float deltaIdle = idleTime - prevIdleTime;
+  float deltaTotal = totalTime - prevTotalTime;
+  float utilization = 100.0 * (1.0 - deltaIdle / deltaTotal);
+  hwConf.cpu = 100 - utilization;
+}
+
+/// MachineStates serialization & deserialization
 NLOHMANN_JSON_SERIALIZE_ENUM(MachineStates, {
     {MachineStates::kInit, "init"},
     {MachineStates::kRestart, "restart"},
@@ -19,7 +103,7 @@ NLOHMANN_JSON_SERIALIZE_ENUM(MachineStates, {
 })
 
 
-// AppStates serialization & deserialization
+/// AppStates serialization & deserialization
 NLOHMANN_JSON_SERIALIZE_ENUM(ApplicationState, {
     {ApplicationState::kInitializing, "init"},
     {ApplicationState::kRunning, "running"},
@@ -115,5 +199,3 @@ void from_json(const json& j, ApplicationManifest& applicationManifest)
 }
 
 }
-
-#endif // JSON_CONVERTERS_HPP
